Add files via upload
This PHP file is responsible for receiving the sensor data via HTTP post and inserting it into a table in a database.
It compares a statically stored API key to authorize incoming sensor data.
It checks if the incoming data is from a new (unregistered) sensor (by checking if the sensor_id exists already). If it is new, it creates a new record in the 'sensor' table and new columns for the new sensor in the temperature, humidity and pressure tables.
If it isn't new then, the data is inserted into the existing columns in the temperature, humidity and pressure table.